[feat] 튜토리얼에서 탄단지 입력 시에 메뉴에서 탄단지 입력처럼 슬라이더가 아닌 텍스트필드로 변경

+++ BalanceEat/Data/DTOs/UserResponseDTO.swift
@@ -20,7 +20,7 @@ struct UserResponseDTO: Codable {
     let smi: Double?
     let fatPercentage: Double?
     let targetWeight: Double
-    let targetCalorie: Int
+    let targetCalorie: Double
     let targetSmi: Double?
     let targetFatPercentage: Double?
     let targetCarbohydrates: Double?
