+++ BalanceEat/Domain/Entities/UserData.swift
@@ -21,7 +21,7 @@ struct UserData {
     let smi: Double?
     let fatPercentage: Double?
     let targetWeight: Double
-    let targetCalorie: Int
+    let targetCalorie: Double
     let targetSmi: Double?
     let targetFatPercentage: Double?
     let targetCarbohydrates: Double?
